instantler/requirements.txt
Django==2.2
django-cors-headers==2.5.2
djangorestframework==3.9.2
psycopg2-binary==2.8
pytz==2018.9
sqlparse==0.3.0
djangorestframework-jwt